refactor(fieldmap): split up FieldmapModals.tsx
